fix(suzuka-full-node): rename task modules
