+++ networks/suzuka/suzuka-full-node/src/tasks/mod.rs
@@ -1,4 +1,4 @@
 //! Modules to separate full node processing into actor-like tasks.
 
-pub mod exec_settle;
-pub mod tx_ingress;
+pub mod execute_settle;
+pub mod transaction_ingress;
